FIR/UAST: clarify parent/source uses in base service APIs
UAST element has a parent linkage if given, and for APIs that convert
the given PsiElement and/or retrieve a sub element, the given UElement
will be uased as a `parent`.

In contrast, UElement given for type conversion is usually an owner of
the PsiElement, and thus it's not quite `parent`. It should rather be
named `source`.

GitOrigin-RevId: d75ae12cb9918fc7120dc7e79a49ea7f8b499a00